Updated to v2.2
This adaption of the early 1.0 version provides faster and more accurate structure and performance as the scraping has been redesigned and compartmentalized as a class. This version also includes the ability to scrape multiple years at a time as well as a finalizing option to have the separate year csv's merged into a completed file. Each of these features is presented as a yes/no prompt following the completion of the scrape of the primary year.
The read and write file path must be customized in this version still. This can be done on lines 21 and 37 respectively.